+++ server/routes/manager_bp.py
@@ -2,19 +2,23 @@ from flask import Blueprint, make_response, jsonify
 from flask_restful import Api, Resource, abort, reqparse
 from flask_bcrypt import Bcrypt
 from flask_marshmallow import Marshmallow
+from flask_jwt_extended import get_jwt_identity
 from serializer import managerSchema
-
+from auth_middleware import manager_required
 from models import Manager, db
 
 manager_bp = Blueprint('manager_bp', __name__)
-ma=Marshmallow(manager_bp)
+ma = Marshmallow(manager_bp)
 bcrypt = Bcrypt()
 api = Api(manager_bp)
 
 post_args = reqparse.RequestParser()
-post_args.add_argument('email', type=str, required=True, help='Email is required')
-post_args.add_argument('password', type=str, required=True, help='Password is required')
-post_args.add_argument('dept_id', type=str, required=True, help='Departmemnt ID  is required')
+post_args.add_argument('email', type=str, required=True,
+                       help='Email is required')
+post_args.add_argument('password', type=str, required=True,
+                       help='Password is required')
+post_args.add_argument('dept_id', type=str, required=True,
+                       help='Departmemnt ID  is required')
 
 patch_args = reqparse.RequestParser()
 patch_args.add_argument('email', type=str)
@@ -29,7 +33,7 @@ class Managers(Resource):
         response = make_response(jsonify(result), 200)
 
         return response
-    
+
     def post(self):
         data = post_args.parse_args()
 
@@ -38,19 +42,22 @@ class Managers(Resource):
         if manager:
             abort(409, detail="Manager with the same email already exists")
         hashed_password = bcrypt.generate_password_hash(data['password'])
-        new_manager = Manager(email=data['email'], password=hashed_password, dept_id=data['dept_id'])
+        new_manager = Manager(
+            email=data['email'], password=hashed_password, dept_id=data['dept_id'])
         db.session.add(new_manager)
         db.session.commit()
 
         result = managerSchema.dump(new_manager)
-        response = make_response(jsonify(result),201)
+        response = make_response(jsonify(result), 201)
 
         return response
-    
-api.add_resource(Managers,'/managers')
 
-class ManagerById(Resource): 
-    def get(self, id): 
+
+api.add_resource(Managers, '/managers')
+
+
+class ManagerById(Resource):
+    def get(self, id):
         single_manager = Manager.query.filter_by(id=id).first()
 
         if not single_manager:
@@ -61,27 +68,38 @@ class ManagerById(Resource):
             response = make_response(jsonify(result), 200)
             return response
 
+    @manager_required()
     def patch(self, id):
+        current_user = get_jwt_identity()
         single_manager = Manager.query.filter_by(id=id).first()
 
         if not single_manager:
             abort(404, detail=f'user with id {id} does not exist')
 
+        if single_manager.id != current_user:
+            abort(401, detail="Unauthorized request")
+
         data = patch_args.parse_args()
         for key, value in data.items():
             if value is None:
                 continue
-            setattr( single_manager, key, value)
+            setattr(single_manager, key, value)
         db.session.commit()
         result = managerSchema.dump(single_manager)
         response = make_response(jsonify(result), 200)
 
         return response
 
+    @manager_required()
     def delete(self, id):
+        current_user = get_jwt_identity()
         manager = Manager.query.filter_by(id=id).first()
         if not manager:
             abort(404, detail=f'manager with id {id} does not exist')
+
+        if manager.id != current_user:
+            abort(401, detail="Unauthorized request")
+
         db.session.delete(manager)
         db.session.commit()
 
@@ -91,4 +109,6 @@ class ManagerById(Resource):
 
         response = make_response(response_body, 200)
         return response
+
+
 api.add_resource(ManagerById, '/managers/<string:id>')
